+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Compatibile with glFusion 1.6.0 or higher.
 - Use LGLIB_ImageUrl to handle image resizing and caching
 - Remove separate configurations for ad and category image locations
 - Change user email to only no or on approval, no notification on rejection
+- Add DB field to hold the category parent tree to save lookups
 
 ## glClassifieds - 1.0.4
 Released 2011-09-02